src/modeling/requirements.txt
boto3==1.26.142
botocore==1.29.142
isort==5.12.0
jmespath==1.0.1
joblib==1.2.0
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==1.26.16
